Fix beachball when parsing multiple directives.
ICU seems to have an issue with certain regexen whereby it'll take a very long
time to do its work on even relatively short strings (tens of characters). The
issue occurs when the regex includes terminal characters (in our case, ;:) but
the string being continually matched does not end in those characters. To get
around it, we naively chop off anything after the last full directive when
receiving data, then we graft it back on to the next data in.